requirements.txt
asgiref==3.4.1
Django==3.2.8
django-countries==7.2.1
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
